+++ tests/test_foreach.py
@@ -5,17 +5,22 @@ from typing import Dict, List, Tuple, Union
 
 import alkymi.recipes
 from alkymi import AlkymiConfig
-from alkymi.alkymi import compute_recipe_status, Status
+from alkymi.alkymi import Status
 import alkymi as alk
 
-# We use this global to avoid altering the hashes of bound functions when the execution count changes
+# We use these globals to avoid altering the hashes of bound functions when any of these change
 execution_counts = {}  # type: Dict[Union[Path, str], int]
+f1 = Path()
+f2 = Path()
+f3 = Path()
 
 
 def test_execution(caplog, tmpdir):
     tmpdir = Path(str(tmpdir))
     caplog.set_level(logging.DEBUG)
     AlkymiConfig.get().cache = False
+
+    global execution_counts, f1, f2, f3
 
     f1 = Path(tmpdir) / "file1.txt"
     f2 = Path(tmpdir) / "file2.txt"
@@ -24,7 +29,6 @@ def test_execution(caplog, tmpdir):
     f2.write_text(f2.stem)
     f3.write_text(f3.stem)
 
-    global execution_counts
     execution_counts = {f1: 0, f2: 0, f3: 0, f1.stem: 0, f2.stem: 0, f3.stem: 0}
 
     def _check_counts(counts: Tuple[int, int, int, int, int, int]):
@@ -55,17 +59,17 @@ def test_execution(caplog, tmpdir):
         execution_counts[file_content] += 1 + extra_count
         return file_content
 
-    assert compute_recipe_status(read_file)[read_file] == Status.NotEvaluatedYet
-    assert compute_recipe_status(change_count)[change_count] == Status.NotEvaluatedYet
+    assert read_file.status() == Status.NotEvaluatedYet
+    assert change_count.status() == Status.NotEvaluatedYet
     change_count.brew()
-    assert compute_recipe_status(read_file)[read_file] == Status.Ok
-    assert compute_recipe_status(change_count)[change_count] == Status.Ok
+    assert read_file.status() == Status.Ok
+    assert change_count.status() == Status.Ok
     _check_counts((1, 0, 0, 1, 0, 0))
 
     args.set_args([f1, f2, f3])
-    assert compute_recipe_status(read_file)[read_file] == Status.MappedInputsDirty
+    assert read_file.status() == Status.MappedInputsDirty
     change_count.brew()
-    assert compute_recipe_status(read_file)[read_file] == Status.Ok
+    assert read_file.status() == Status.Ok
     _check_counts((1, 1, 1, 1, 1, 1))
 
     args.set_args([f3, f2])
@@ -82,17 +86,20 @@ def test_execution(caplog, tmpdir):
 
     # Test that changing an ingredient forces reevaluation of all foreach inputs
     args.set_args([f1, f2, f3])
-    assert compute_recipe_status(change_count)[change_count] == Status.MappedInputsDirty
+    assert change_count.status() == Status.MappedInputsDirty
     change_count.brew()
-    assert compute_recipe_status(change_count)[change_count] == Status.Ok
+    assert change_count.status() == Status.Ok
     _check_counts((3, 2, 2, 3, 2, 2))
 
     # This should cause a reevaluation of everything (+1 to all counts) and then add 10 from this arg
     extra_count_arg.set_args(10)
-    assert compute_recipe_status(change_count)[change_count] == Status.IngredientDirty
+    assert change_count.status() == Status.IngredientDirty
     change_count.brew()
-    assert compute_recipe_status(change_count)[change_count] == Status.Ok
+    assert change_count.status() == Status.Ok
     _check_counts((3, 2, 2, 14, 13, 13))
+
+
+execution_counts_list = []  # type: List[int]
 
 
 def test_lists(caplog):
@@ -102,20 +109,21 @@ def test_lists(caplog):
     caplog.set_level(logging.DEBUG)
     AlkymiConfig.get().cache = False
 
-    execution_counts = [0] * 5  # type: List[int]
+    global execution_counts_list
+    execution_counts_list = [0] * 5
     args = alkymi.recipes.args([0])
 
     def _check_counts(counts: Tuple[int, int, int, int, int]):
-        assert execution_counts[0] == counts[0]
-        assert execution_counts[1] == counts[1]
-        assert execution_counts[2] == counts[2]
-        assert execution_counts[3] == counts[3]
-        assert execution_counts[4] == counts[4]
+        assert execution_counts_list[0] == counts[0]
+        assert execution_counts_list[1] == counts[1]
+        assert execution_counts_list[2] == counts[2]
+        assert execution_counts_list[3] == counts[3]
+        assert execution_counts_list[4] == counts[4]
 
     @alk.foreach(args.recipe)
     def record_execution(idx: int) -> int:
-        execution_counts[idx] += 1
-        return execution_counts[idx]
+        execution_counts_list[idx] += 1
+        return execution_counts_list[idx]
 
     # Initial brew should only increment first id
     record_execution.brew()